requirements.txt
pytest
pytest-cov
behave
fastapi
uvicorn[standard]
jinja2
python-multipart
email-validator
sqlalchemy[asyncio]>=2.0
alembic
psycopg[binary]
python-dotenv
aiosqlite
passlib[argon2]
python-jose[cryptography]